Development/code-archive/anya_sammy work/phone-camera/1510_merged_code_simulation/requirements.txt
numpy>=1.24
pyserial>=3.5
pyttsx3>=2.90
pynput>=1.7.6
opencv-python>=4.8
sounddevice


# if no GUI(Server/CI) - for headless opencv
# change to - opencv-python-headless>=4.8
# if GUI(Laptop/Desktop) - for opencv with GUI
# change to - opencv-python>=4.8